+++ src/main/scala/de/uni_luebeck/isp/tessla/Compiler.scala
@@ -3,19 +3,30 @@ package de.uni_luebeck.isp.tessla
 import org.antlr.v4.runtime.CharStream
 
 object Compiler {
-  def instantiatePipeline(unit: Option[TimeUnit]): TranslationPhase[CharStream, TesslaCore.Specification] = {
-    TesslaParser
+  case class Options(
+    timeUnitString: Option[String],
+    includeResolver: String => Option[CharStream],
+    stdlibIncludeResolver: String => Option[CharStream],
+    stdlibPath: String,
+    currySignalLift: Boolean
+  ) {
+    lazy val timeUnit = timeUnitString.map(TimeUnit.fromString(_, Location.option("timeunit")))
+  }
+
+  def instantiatePipeline(options: Options): TranslationPhase[CharStream, TesslaCore.Specification] = {
+    new TesslaParser.WithIncludes(options.includeResolver)
+      .andThen(new StdlibIncluder(options.stdlibIncludeResolver, options.stdlibPath))
       .andThen(TesslaSyntaxToTessla)
       .andThen(Flattener)
       .andThen(TypeChecker)
-      .andThen(new ConstantEvaluator(unit))
+      .andThen(new ConstantEvaluator(options.timeUnit))
       .andThen(CycleDetection)
-      .andThen(CurrySignalLift)
+      .andThen(new EnableIf[TesslaCore.Specification](options.currySignalLift, CurrySignalLift))
       .andThen(RemoveUnusedDefinitions)
   }
 
-  def compile(src: CharStream, unit: Option[TimeUnit]): TranslationPhase.Result[TesslaCore.Specification] = {
-    instantiatePipeline(unit).translate(src)
+  def compile(src: CharStream, options: Options): TranslationPhase.Result[TesslaCore.Specification] = {
+    instantiatePipeline(options).translate(src)
   }
 
   class Printer[T] extends TranslationPhase[T, T] {
@@ -26,8 +37,13 @@ object Compiler {
     }
   }
 
-  def compile(src: CharStream, timeUnit: Option[String])(implicit dummy:DummyImplicit): TranslationPhase.Result[TesslaCore.Specification] = {
-    val parsedTimeUnit = timeUnit.map(TimeUnit.fromString(_, Location.option("timeunit")))
-    compile(src, parsedTimeUnit)
+  class EnableIf[T](condition: Boolean, phase: TranslationPhase[T,T]) extends TranslationPhase[T, T] {
+    override def translate(spec: T) = {
+      if (condition) {
+        phase.translate(spec)
+      } else {
+        TranslationPhase.Success(spec, Seq())
+      }
+    }
   }
 }