+++ src/main/scala/de/uni_luebeck/isp/tessla/interpreter/Interpreter.scala
@@ -14,7 +14,7 @@ class Interpreter(val spec: TesslaCore.Specification) extends Specification {
   type Env = Map[TesslaCore.Identifier, TesslaCore.ValueOrError]
 
   lazy val defs: Map[TesslaCore.Identifier, Lazy[Stream]] = spec.streams.map { stream =>
-    stream.id -> Lazy(eval(stream.expression))
+    stream.id -> Lazy(eval(stream))
   }.toMap
 
   lazy val outStreams: Seq[(Option[String], Stream, TesslaCore.Type)] = spec.outStreams.map { os =>
@@ -26,10 +26,10 @@ class Interpreter(val spec: TesslaCore.Specification) extends Specification {
   }
 
   private def evalStream(arg: TesslaCore.StreamRef): Stream = arg match {
-    case TesslaCore.Stream(id, loc) =>
-      defs.getOrElse(id, throw InternalError(s"Couldn't find stream named $id", loc)).get
-    case TesslaCore.InputStream(name, loc) =>
-      inStreams.getOrElse(name, throw InternalError(s"Couldn't find input stream named $name", loc))._1
+    case s: TesslaCore.Stream =>
+      defs.getOrElse(s.id, throw InternalError(s"Couldn't find stream named ${s.id}", s.loc)).get
+    case i: TesslaCore.InputStream =>
+      inStreams.getOrElse(i.name, throw InternalError(s"Couldn't find input stream named ${i.name}", i.loc))._1
     case _: TesslaCore.Nil => nil
   }
 
@@ -37,15 +37,6 @@ class Interpreter(val spec: TesslaCore.Specification) extends Specification {
     lift(Seq(stream1, stream2)) {
       case Seq(Some(value1), _) => Some(value1)
       case Seq(None, value2Opt) => value2Opt
-    }
-  }
-
-  def filter(events: Stream, condition: Stream): Stream = {
-    val latestCondition = merge(condition, last(condition, events))
-    lift(Seq(events, latestCondition)) {
-      case Seq(value, Some(TesslaCore.BoolValue(true, _))) =>
-        value
-      case _ => None
     }
   }
 
@@ -65,41 +56,7 @@ class Interpreter(val spec: TesslaCore.Specification) extends Specification {
     }
   }
 
-  // def foldTotal[A](x: Events[A], f: (A, A) => A) = {
-  //   def y: Events[A] := merge(lift(last(y, x), x, (a: Option[A], b: Option[A]) =>
-  //     if isNone(a) || isNone(b) then None[A]
-  //     else Some(f(getSome(a), getSome(b)))), x)
-  //   y
-  // }
-  def foldTotal(x: Stream, f: (TesslaCore.Value, TesslaCore.Value) => TesslaCore.ValueOrError): Stream = {
-    lazy val y: Stream = merge(lift(Seq(last(y, x), x)) {
-      case Seq(Some(a), Some(b)) =>
-        Some(a.mapValue(aValue => b.mapValue(bValue => f(aValue, bValue))))
-      case _ => None
-    }, x)
-    y
-  }
-
-  // def foldTotalWithInit[A,B](x: Events[A], init: B, f: (B, A) => B) = {
-  //   def y: Events[B] := default(lift(last(y, x), x, (a: Option[B], b: Option[A]) =>
-  //     if isNone(b) then None[B]
-  //     else if isNone(a) then Some(f(init, getSome(b)))
-  //     else Some(f(getSome(a), getSome(b)))), init)
-  //   y
-  // }
-  def foldTotal(x: Stream, f: (TesslaCore.Value, TesslaCore.Value) => TesslaCore.ValueOrError, init: TesslaCore.ValueOrError): Stream = {
-    lazy val y: Stream = lift(Seq(last(y, x), x)) {
-      case Seq(None, Some(b)) =>
-        Some(init.mapValue(aValue => b.mapValue(bValue => f(aValue, bValue))))
-      case Seq(Some(a), Some(b)) =>
-        Some(a.mapValue(aValue => b.mapValue(bValue => f(aValue, bValue))))
-      case _ => None
-    }.default(init)
-    y
-  }
-
-
-  private def eval(exp: TesslaCore.Expression): Stream = exp match {
+  private def eval(sd: TesslaCore.StreamDescription): Stream = sd.expression match {
     case TesslaCore.SignalLift(op, argStreams, loc) =>
       if (argStreams.isEmpty) {
         throw Errors.InternalError("Lift without arguments should be impossible", loc)
@@ -108,7 +65,7 @@ class Interpreter(val spec: TesslaCore.Specification) extends Specification {
         val args = arguments.zip(argStreams).map {
           case (arg, stream) => arg.mapValue(_.withLoc(stream.loc))
         }
-        Some(Evaluator.evalPrimitiveOperator(op, args, exp.loc))
+        Some(Evaluator.evalPrimitiveOperator(op, args, sd.typ.elementType, sd.loc))
       }
     case TesslaCore.Lift(f, argStreams, loc) =>
       if (argStreams.isEmpty) {
@@ -116,12 +73,13 @@ class Interpreter(val spec: TesslaCore.Specification) extends Specification {
       }
       lift(argStreams.map(evalStream)) { arguments =>
         val args = arguments.zip(argStreams).map {
-          case (Some(arg), stream) => arg.mapValue(a => TesslaCore.TesslaOption(Some(a), stream.loc))
-          case (None, stream) => TesslaCore.TesslaOption(None, stream.loc)
+          case (Some(arg), stream) => TesslaCore.TesslaOption(Some(arg), stream.typ.elementType, stream.loc)
+          case (None, stream) => TesslaCore.TesslaOption(None, sd.typ.elementType, stream.loc)
         }
         val closure = TesslaCore.Closure(f, Map(), loc)
-        Evaluator.evalApplication(closure, args, loc) match {
+        Evaluator.evalApplication(closure, args, sd.typ.elementType, loc) match {
           case to: TesslaCore.TesslaOption => to.value
+          case err: TesslaCore.Error => Some(err)
           case other =>
             throw InternalError(s"Used lift on non-option function (return value: $other) - should have been caught by type checker")
         }
@@ -130,40 +88,14 @@ class Interpreter(val spec: TesslaCore.Specification) extends Specification {
       evalStream(values).default(defaultValue)
     case TesslaCore.DefaultFrom(values, defaults, _) =>
       evalStream(values).default(evalStream(defaults))
-    case TesslaCore.Const(value, stream, _) =>
-      slift(Seq(evalStream(stream))) { _ =>
-        Some(value)
-      }
     case TesslaCore.Last(values, clock, _) =>
       last(evalStream(values), evalStream(clock))
-    case TesslaCore.DelayedLast(values, delays, _) =>
-      delayedLast(evalStream(delays), evalStream(values))
     case TesslaCore.Delay(delays, resets, _) =>
       delay(evalStream(delays), evalStream(resets))
     case TesslaCore.Time(values, loc) =>
       evalStream(values).time(loc)
-    case TesslaCore.StdLibCount(values, loc) =>
-      foldTotal(evalStream(values),
-        (a,_) => TesslaCore.IntValue(getInt(a) + 1, loc),
-        TesslaCore.IntValue(0, loc))
-    case TesslaCore.StdLibSum(values, loc) =>
-      foldTotal(evalStream(values),
-        (a,b) => TesslaCore.IntValue(getInt(b) + getInt(a), loc),
-        TesslaCore.IntValue(0, loc))
-    case TesslaCore.StdLibMinimum(values, loc) =>
-      foldTotal(evalStream(values),
-        (a,b) => TesslaCore.IntValue(getInt(a) min getInt(b), loc))
-    case TesslaCore.StdLibMaximum(values, loc) =>
-      foldTotal(evalStream(values),
-        (a,b) => TesslaCore.IntValue(getInt(a) max getInt(b), loc))
-    case TesslaCore.Merge(arg1, arg2, _) =>
-      val stream1 = evalStream(arg1)
-      val stream2 = evalStream(arg2)
-      merge(stream1, stream2)
-    case TesslaCore.Filter(eventsArg, conditionArg, _) =>
-      val eventsStream = evalStream(eventsArg)
-      val conditionStream = evalStream(conditionArg)
-      filter(eventsStream, conditionStream)
+    case customCall: TesslaCore.CustomBuiltInCall =>
+      throw InternalError(s"The interpreter does not support any custom built-ins (${customCall.name})")
   }
 
   def getInt(value: TesslaCore.Value): BigInt = value match {